requirements-infra.txt
cdktf
cdktf-cdktf-provider-azurerm
cdktf-cdktf-provider-azuread
cdktf-cdktf-provider-github
pydantic-settings